Move vision feature memoizer to module
