+++ mlx_engine/model_kit/model_kit.py
@@ -262,8 +262,6 @@ class ModelKit:
         return False
 
     def shutdown(self) -> None:
-        if self.vision_add_on is not None:
-            self.vision_add_on.clear_feature_cache()
         self._shutdown.set()
 
     def is_shutdown(self) -> None:
