+++ mlx_engine/model_kit/vision_add_ons/gemma3n.py
@@ -139,7 +139,7 @@ class Gemma3nVisionAddOn(BaseVisionAddOn):
                 self.embed_vision,
             )
 
-        image_features = self.vision_feature_cache.get_or_compute(
+        image_features = self._vision_feature_memoizer.get_or_compute(
             images_b64, max_size, compute_image_features
         )
