+++ mlx_engine/model_kit/vision_add_ons/gemma4.py
@@ -101,7 +101,7 @@ class Gemma4VisionAddOn(BaseVisionAddOn):
                 input_embeddings.dtype
             )
 
-        image_features = self.vision_feature_cache.get_or_compute(
+        image_features = self._vision_feature_memoizer.get_or_compute(
             images_b64, max_size, compute_image_features
         ).astype(input_embeddings.dtype)
